refactor: move spinner logic to hook
